refactor: creation d'une wrapper jwt dédié
